requirements.txt
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
psutil==5.4.3
py4j==0.10.6
pyenchant==2.0.0
pyspark==2.3.0
python-dateutil==2.7.2
pytz==2018.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
